bot/requirements.txt
asyncpg
claptcha
discord.py
redis
tabulate
pyyaml